Discord_Project/city-of-tucson-ai-core/learning/langchain_discord_bot/requirements.txt
discord.py
langchain
openai
chromadb
python-dotenv